perf: use the faster binarize implementation
